i2/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.0
google-generativeai==0.3.2
gunicorn==21.2.0